rlist: introduce rlist_add_tail_entry_sorted
Add an entry to an ascending sorted list. Scan is started from
the tail. Applicable when new elements are usually already bigger
than all other ones, since insertion into a sorted list is O(N).

Necessary for SWIM implementation, where a list is stored of
members, waiting for an ACK. Sometimes there can be inserted an
indirect ACK, which deadline could be bigger, than deadlines
of next direct ACKs. So new elements are inserted either into the
tail, or almost into the tail.

Order allows to stop checking next elements for an unacknowledged
ping when on a next one deadline is not exceeded yet.

Needed for #3234